Add support for 'alias' attribute on Appfwk Columns
In AR11, many key values are bare enums,
with the more descriptive values placed as value columns, e.g.:

Key: dns.is_success		- true/false
Value dns.is_success_name  - Success/Failure

Often the more informative column is the "name" version,
but as a 'value' column it can't be retrieved without the
corresponding 'key' column.

This commit adds an attribute to the AppResponse Column so the
alias column gets pulled when the query is made, then swapped in
for the less readable column before the results are returned.